requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.5.0
ansi2html==1.8.0
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.12.2
bleach==6.0.0
blessed==1.20.0
blinker==1.6.2
bottleneck==1.3.7
brotlipy==0.7.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.0.4
click==8.1.3
cloudpickle==2.0.0
colorama==0.4.6
croniter==1.4.1
cryptography==39.0.1
cycler==0.11.0
dash==2.9.3
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dateutils==0.6.12
debugpy==1.5.1
decorator==5.1.1
deepdiff==6.4.1
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
docker-pycreds==0.4.0
entrypoints==0.4
exceptiongroup==1.0.4
executing==1.2.0
faker==19.6.1
fastapi==0.103.1
fastjsonschema==2.16.3
favicon==0.7.0
filelock==3.12.0
finetuning-scheduler==2.0.7
flask==2.3.2
flit-core==3.8.0
frozenlist==1.4.0
fsspec==2023.4.0
gdown==4.7.1
gitdb==4.0.10
gitpython==3.1.31
gmpy2==2.1.2
greenlet==2.0.1
h11==0.14.0
htbuilder==0.6.2
identify==2.5.23
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
iniconfig==1.1.1
inquirer==3.1.3
ipykernel==6.15.0
ipython==8.13.2
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.2
jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter-client==7.4.9
jupyter-core==4.12.0
jupyter-dash==0.4.2
jupyter-server==1.23.6
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
lightning==2.0.7
lightning-cloud==0.5.37
lightning-utilities==0.8.0
lxml==4.9.3
markdown==3.4.4
markdown-it-py==3.0.0
markdownlit==0.0.7
markupsafe==2.1.1
matplotlib==3.4.3
matplotlib-inline==0.1.6
mdurl==0.1.2
memory-profiler==0.61.0
mistune==2.0.5
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
molplotly==1.1.7
more-itertools==10.1.0
mpmath==1.2.1
multidict==6.0.4
multiprocess==0.70.14
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==2.8.4
nodeenv==1.7.0
notebook==6.5.4
notebook-shim==0.2.3
numexpr==2.8.4
numpy==1.24.3
ordered-set==4.1.0
packaging==23.1
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathos==0.3.0
pathtools==0.1.2
pexpect==4.8.0
pickle-mixin==1.0.2
pickleshare==0.7.5
pillow==9.4.0
pip==23.0.1
pkgutil-resolve-name==1.3.10
platformdirs==3.5.0
plotly==5.14.1
pluggy==1.0.0
pooch==1.4.0
pox==0.3.2
ppft==1.7.6.6
pre-commit==3.3.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.23.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pycairo==1.23.0
pycparser==2.21
pydantic==2.1.1
pydantic-core==2.4.0
pydeck==0.8.0
pygments==2.15.1
pyjwt==2.8.0
pymdown-extensions==10.3
pympler==1.0.1
pyopenssl==23.0.0
pyparsing==3.0.9
pyrsistent==0.19.3
pysocks==1.7.1
pytest==7.3.1
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pyyaml==6.0
pyzmq==23.2.0
rdkit-pypi==2022.9.5
readchar==4.0.5
reportlab==3.6.13
requests==2.29.0
retrying==1.3.4
rich==13.5.2
scikit-learn==1.0.2
scipy==1.10.1
send2trash==1.8.2
sentry-sdk==1.22.2
setproctitle==1.3.2
setuptools==66.0.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlalchemy==1.4.46
st-annotated-text==4.0.1
stack-data==0.6.2
starlette==0.27.0
starsessions==1.3.0
streamlit==1.26.0
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-embedcode==0.1.2
streamlit-extras==0.3.2
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
syba==1.0.2a0
sympy==1.11.1
tenacity==8.2.2
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.0
torch-geometric==2.3.0
torch-scatter==2.1.1
torchaudio==2.0.0
torchmetrics==0.11.4
torchvision==0.15.0
tornado==6.3
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing-extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.15
uvicorn==0.23.2
validators==0.22.0
virtualenv==20.23.0
wandb==0.15.2
watchdog==3.0.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.3
werkzeug==2.3.4
wheel==0.38.4
yarl==1.9.2
zipp==3.15.0
